LPS-104599 Reorganizing the upgrade processes to be able to backport it to 7.2.x
