+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-service/src/main/java/com/liferay/dynamic/data/mapping/internal/upgrade/DDMServiceUpgrade.java
@@ -267,18 +267,18 @@ public class DDMServiceUpgrade implements UpgradeStepRegistrator {
 		registry.register("3.2.0", "3.2.1", new DummyUpgradeStep());
 
 		registry.register(
-			"3.2.1", "3.3.0",
+			"3.2.1", "3.2.2",
+			new com.liferay.dynamic.data.mapping.internal.upgrade.v3_2_2.
+				UpgradeDDMFormFieldValidation(_jsonFactory));
+
+		registry.register(
+			"3.2.2", "3.3.0",
 			new UpgradeCTModel(
 				"DDMStructure", "DDMStructureVersion", "DDMTemplate",
 				"DDMTemplateVersion"));
 
 		registry.register(
-			"3.3.0", "3.3.1",
-			new com.liferay.dynamic.data.mapping.internal.upgrade.v3_3_1.
-				UpgradeDDMFormFieldValidation(_jsonFactory));
-
-		registry.register(
-			"3.3.1", "3.4.0",
+			"3.3.0", "3.4.0",
 			new UpgradeCTModel("DDMStructureLink", "DDMTemplateLink"));
 	}
